implemented changes as per review suggestions
