+++ src/embedded/embeddedClickEvents.ts
@@ -1,11 +1,17 @@
 import {
-  EmbeddedManager,
   EmbeddedMessageElementsButton,
+  IterableAction,
+  IterableActionRunner,
+  IterableActionSource,
   IterableEmbeddedMessage
 } from '../embedded';
 import { ErrorHandler, trackEmbeddedClick } from '..';
+import {
+  URL_SCHEME_ACTION,
+  URL_SCHEME_ITBL,
+  URL_SCHEME_OPEN
+} from 'src/constants';
 
-const embeddedManager = new EmbeddedManager();
 export const handleElementClick = (
   message: IterableEmbeddedMessage,
   errorCallback?: ErrorHandler
@@ -14,12 +20,18 @@ export const handleElementClick = (
     message?.elements?.defaultAction?.data?.trim() ||
     message?.elements?.defaultAction?.type ||
     null;
-  embeddedManager.handleEmbeddedClick(message, null, clickedUrl);
+  handleEmbeddedClick(clickedUrl);
   trackEmbeddedClick({
     messageId: message.metadata.messageId,
     buttonIdentifier: '',
-    clickedUrl: clickedUrl ? clickedUrl : '',
-    errorCallback
+    clickedUrl: clickedUrl ? clickedUrl : ''
+  }).catch((error) => {
+    if (errorCallback) {
+      errorCallback({
+        ...error?.response?.data,
+        statusCode: error?.response?.status
+      });
+    }
   });
 };
 
@@ -28,14 +40,19 @@ export const handleButtonClick = (
   message: IterableEmbeddedMessage,
   errorCallback?: ErrorHandler
 ) => {
-  const embeddedManager = new EmbeddedManager();
   const clickedUrl = button?.action?.data?.trim() || button?.action?.type || '';
-  embeddedManager.handleEmbeddedClick(message, button?.id || null, clickedUrl);
+  handleEmbeddedClick(clickedUrl);
   trackEmbeddedClick({
     messageId: message.metadata.messageId,
     buttonIdentifier: button?.id || '',
-    clickedUrl,
-    errorCallback
+    clickedUrl
+  }).catch((error) => {
+    if (errorCallback) {
+      errorCallback({
+        ...error?.response?.data,
+        statusCode: error?.response?.status
+      });
+    }
   });
 };
 
@@ -57,4 +74,33 @@ export const addButtonClickEvent = (
       errorCallback
     );
   });
+};
+
+const handleEmbeddedClick = (clickedUrl: string | null) => {
+  if (clickedUrl && clickedUrl.trim() !== '') {
+    let actionType: string;
+    let actionName: string;
+
+    if (clickedUrl.startsWith(URL_SCHEME_ACTION)) {
+      actionName = '';
+      actionType = clickedUrl;
+    } else if (clickedUrl.startsWith(URL_SCHEME_ITBL)) {
+      actionName = '';
+      actionType = clickedUrl.replace(URL_SCHEME_ITBL, '');
+    } else {
+      actionType = URL_SCHEME_OPEN;
+      actionName = clickedUrl;
+    }
+
+    const iterableAction: IterableAction = {
+      type: actionType,
+      data: actionName
+    };
+
+    IterableActionRunner.executeAction(
+      null,
+      iterableAction,
+      IterableActionSource.EMBEDDED
+    );
+  }
 };
