+++ src/events/embedded/types.ts
@@ -1,5 +1,3 @@
-import { ErrorHandler } from 'src/types';
-
 export interface IEmbeddedMessageElementsButton {
   id: string;
   title: string;
@@ -97,7 +95,6 @@ export interface EmbeddedTrackClick {
   buttonIdentifier: string;
   clickedUrl: string;
   appPackageName?: string;
-  errorCallback?: ErrorHandler;
 }
 
 export interface IEmbeddedImpressionData {
