+++ src/types.ts
@@ -35,14 +35,6 @@ export interface IterableResponse {
   params?: null | Record<string, any>;
 }
 
-export interface ClientError extends IterableResponse {
-  clientErrors: {
-    error: string;
-    field?: string;
-  }[];
-  statusCode?: number;
-}
-
 export interface ErrorHandler {
-  (error: ClientError): void;
+  (error: any): void;
 }
